+++ CHANGELOG.md
@@ -4,6 +4,14 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ## [Unreleased]
 
+### Breaking changes
+
+- `saveSolidDatasetInContainer`, `saveFileInContainer` and `createContainerInContainer` now return
+  `null` if the newly-created Resource is not actually readable by the current user. Please update
+  your code with a `null` check on the return value before writing to the Resource again, or read
+  the input value if you want to inspect the sent data without Read permissions to the saved
+  Resource.
+
 ### New features
 
 - `deleteSolidDataset` and `deleteContainer`: two functions that allow you to delete a SolidDataset